fix(parser): normalize ParseErr and add doctests
